fix preprocess

+++ include/nncase/transforms/neutral/pre_process_setting.h
@@ -20,7 +20,7 @@ namespace nncase::ir::transforms
 class NNCASE_API pre_process_transform : public graph_pass
 {
 public:
-    pre_process_transform(std::vector<float> mean, std::vector<float> scale, std::vector<float> input_range, std::vector<float> input_shape, std::string image_format, std::string input_layout, std::string input_type, std::string quant_type, std::string real_layout) noexcept
+    pre_process_transform(std::vector<float> mean, std::vector<float> scale, std::vector<float> input_range, std::vector<int32_t> input_shape, std::string image_format, std::string input_layout, std::string input_type, std::string quant_type, std::string real_layout) noexcept
         : means_(std::move(mean)), scales_(std::move(scale)), input_range_(input_range), input_shape_(input_shape), image_format_(image_format), input_layout_(input_layout), input_type_(input_type), quant_type_(quant_type), real_layout_(real_layout) {};
     using graph_pass::graph_pass;
 
@@ -31,7 +31,7 @@ private:
     std::vector<float> means_;
     std::vector<float> scales_;
     std::vector<float> input_range_;
-    std::vector<float> input_shape_;
+    std::vector<int32_t> input_shape_;
     std::string image_format_;
     std::string input_layout_;
     std::string input_type_;
